+++ frontend/app/chat/components/ChatMessages/ChatInput/MicButton.tsx
@@ -1,7 +1,7 @@
 "use client";
 import Button from "@/lib/components/ui/Button";
+import { useSpeech } from "@/lib/context/ChatsProvider/hooks/useSpeech";
 import { MdMic, MdMicOff } from "react-icons/md";
-import { useSpeech } from "../../../ChatsProvider/hooks/useSpeech";
 
 export function MicButton() {
   const { isListening, speechSupported, startListening } = useSpeech();
