+++ frontend/app/chat/components/ChatMessages/index.tsx
@@ -1,7 +1,7 @@
 "use client";
 import Card from "@/lib/components/ui/Card";
+import useChatsContext from "@/lib/context/ChatsProvider/hooks/useChatsContext";
 import { FC, useEffect, useRef } from "react";
-import useChatsContext from "../../ChatsProvider/hooks/useChatsContext";
 import ChatMessage from "./ChatMessage";
 
 export const ChatMessages: FC = () => {
